+++ src/main/java/function/coverage/comparison/ExonCleanLinear.java
@@ -165,7 +165,7 @@ public class ExonCleanLinear {
                 covDiff = MathManager.abs(caseAvg, ctrlAvg);
             }
 
-            sb.append(FormatManager.getDouble(covDiff)).append(",");
+            sb.append(FormatManager.getFloat(covDiff)).append(",");
             sb.append(geneSize);
             return sb.toString();
         } else {
